// src/app/models/vendedor/shared/vendedor.ts
export class Vendedor {
    id: number;
    codigo: number;
    nome: string;
    rg: string;
    cpf: string;
    dataNascimento: Date;
    email: string;
    telefone: string;       
    operadora: string;
    celular: string;
    operadoraII: string;
    celularII: string;
    municipio: string;
    sigla: string;
    cep: string;
    bairro: string;
    endereco: string;
    numero: string;
    observacao: string;
    salario: string;
    pedidoInicial: string;
    pedidoFinal: string;
    registroInicial: string;
    registroFinal: string;
    cliente: [{}];
}